# Add Json configuration flag to allow comments