--- a/formats/json-tests/jvmTest/resources/spec_cases/listing_comments.txt
+++ b/formats/json-tests/jvmTest/resources/spec_cases/listing_comments.txt
@@ -1,0 +1,4 @@
+// Non-spec inputs that we accept with allowComments = true setting:
+n_object_trailing_comment.json
+n_object_trailing_comment_slash_open.json
+n_structure_object_with_comment.json