--- a/formats/json/commonMain/src/kotlinx/serialization/json/internal/lexer/ReaderJsonLexer.kt
+++ b/formats/json/commonMain/src/kotlinx/serialization/json/internal/lexer/ReaderJsonLexer.kt
@@ -4,6 +4,9 @@
 
 package kotlinx.serialization.json.internal
 
+import kotlinx.serialization.*
+import kotlinx.serialization.json.*
+
 internal const val BATCH_SIZE: Int = 16 * 1024
 private const val DEFAULT_THRESHOLD = 128
 
@@ -32,9 +35,13 @@ internal class ArrayAsSequence(internal val buffer: CharArray) : CharSequence {
     override fun toString(): String = substring(0, length)
 }
 
-internal class ReaderJsonLexer(
-    private val reader: InternalJsonReader,
-    private val buffer: CharArray = CharArrayPoolBatchSize.take()
+@OptIn(ExperimentalSerializationApi::class)
+internal fun ReaderJsonLexer(json: Json, reader: InternalJsonReader, buffer: CharArray= CharArrayPoolBatchSize.take()) =
+    if (json.configuration.allowComments) ReaderJsonLexerWithComments(reader, buffer) else ReaderJsonLexer(reader, buffer)
+
+internal open class ReaderJsonLexer(
+    val reader: InternalJsonReader,
+    val buffer: CharArray = CharArrayPoolBatchSize.take()
 ) : AbstractJsonLexer() {
     private var threshold: Int = DEFAULT_THRESHOLD // chars
 
@@ -42,16 +49,6 @@ internal class ReaderJsonLexer(
 
     init {
         preload(0)
-    }
-
-    override fun tryConsumeComma(): Boolean {
-        val current = skipWhitespaces()
-        if (current >= source.length || current == -1) return false
-        if (source[current] == ',') {
-            ++currentPosition
-            return true
-        }
-        return false
     }
 
     override fun canConsumeValue(): Boolean {
@@ -119,6 +116,41 @@ internal class ReaderJsonLexer(
         }
         currentPosition = cpos
         return TC_EOF
+    }
+
+    override fun consumeNextToken(expected: Char) {
+        ensureHaveChars()
+        val source = source
+        var cpos = currentPosition
+        while (true) {
+            cpos = prefetchOrEof(cpos)
+            if (cpos == -1) break // could be inline function but KT-1436
+            val c = source[cpos++]
+            if (c == ' ' || c == '\n' || c == '\r' || c == '\t') continue
+            currentPosition = cpos
+            if (c == expected) return
+            unexpectedToken(expected)
+        }
+        currentPosition = cpos
+        unexpectedToken(expected) // EOF
+    }
+
+    override fun skipWhitespaces(): Int {
+        var current = currentPosition
+        // Skip whitespaces
+        while (true) {
+            current = prefetchOrEof(current)
+            if (current == -1) break
+            val c = source[current]
+            // Faster than char2TokenClass actually
+            if (c == ' ' || c == '\n' || c == '\r' || c == '\t') {
+                ++current
+            } else {
+                break
+            }
+        }
+        currentPosition = current
+        return current
     }
 
     override fun ensureHaveChars() {