… benchmark/src/jmh/kotlin/kotlinx/benchmarks/json/TwitterFeedCommentsBenchmark.kt
… benchmark/src/jmh/kotlin/kotlinx/benchmarks/json/TwitterFeedStreamBenchmark.kt
… formats/json-tests/jvmTest/resources/spec_cases/listing.txt
… formats/json-tests/jvmTest/src/kotlinx/serialization/json/SpecConformanceTest.kt
… formats/json/commonMain/src/kotlinx/serialization/json/internal/lexer/StringJsonLexer.kt