in C/Java style for both string and stream parser.

This flag, together with allowTrailingCommas and isLenient, will help to cover most use cases for Json5, such as configuration files.

Fixes #2221
Fixes #797