@shanshin 2024-03-15
Could you also please show the benchmark results?
I am interested in the results before the changes made, after them, as well as after the changes, but if you set `allowComments = true` for all available benchmark tests (without changing of JSON strings).