@sandwwraith 2024-03-18
> after the changes, but if you set allowComments = true for all available benchmark tests

I believe `TwitterFeedCommentsBenchmark.decodeMicroTwitterCommentSupport` does exactly this, as it decodes data without comments in `allowComments = true` mode. Although I can re-check with every test if you insist.