@sandwwraith 2024-03-18
For other benchmarks

Before (`dev` branch):
```
Benchmark                                                    Mode  Cnt     Score    Error  Units
TwitterFeedBenchmark.decodeMicroTwitter                     thrpt   30  1036.555 ±  5.558  ops/s
TwitterFeedBenchmark.decodeMicroTwitterNoAltNames           thrpt   30  1097.467 ±  5.286  ops/s
TwitterFeedBenchmark.decodeMicroTwitterWithNamingStrategy   thrpt   30   971.212 ± 12.652  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterJacksonStream  thrpt   30   605.712 ±  6.738  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterReadText       thrpt   30   484.507 ±  5.646  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterStream         thrpt   30   496.026 ±  1.806  ops/s
```

After:
```
Benchmark                                                             Mode  Cnt     Score    Error  Units
TwitterFeedBenchmark.decodeMicroTwitter                              thrpt   30  1016.684 ±  3.544  ops/s
TwitterFeedBenchmark.decodeMicroTwitterNoAltNames                    thrpt   30  1097.850 ±  4.275  ops/s
TwitterFeedBenchmark.decodeMicroTwitterWithNamingStrategy            thrpt   30   996.990 ± 22.545  ops/s
TwitterFeedCommentsBenchmark.decodeMicroTwitter                      thrpt   30  1057.715 ±  5.114  ops/s
TwitterFeedCommentsBenchmark.decodeMicroTwitterCommentInData         thrpt   30   860.290 ±  9.344  ops/s
TwitterFeedCommentsBenchmark.decodeMicroTwitterCommentInDataStream   thrpt   30   550.861 ± 14.833  ops/s
TwitterFeedCommentsBenchmark.decodeMicroTwitterCommentSupport        thrpt   30   908.675 ± 15.589  ops/s
TwitterFeedCommentsBenchmark.decodeMicroTwitterCommentSupportStream  thrpt   30   649.064 ±  5.674  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterJacksonStream           thrpt   30   602.225 ±  1.965  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterReadText                thrpt   30   465.921 ±  1.146  ops/s
TwitterFeedStreamBenchmark.decodeMicroTwitterStream                  thrpt   30   545.131 ± 19.263  ops/s
```